Added streams functionality to CUDA generator
Code uses the liveness information from the liveVarsAnalyser and creates
a dependency graph. It adds every new kernel call to a new stream as
required. It creates dependencies between the streams using events. If
any non-forAll statement is encountered the dependency graph creation is
stopped and a cudaDevideSynchronize() is used to sync before running any
other statement.